+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 0.3.2
 
 - Hide menu when no children menu.
+- Add custom filters.
 
 ### 0.3.1